Fix case statement to use ::Numeric and ::String
